backend/requirements.txt
# ─── Web framework ───────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# ─── HTTP client ─────────────────────────────────────────────
httpx==0.27.2

# ─── Supabase ────────────────────────────────────────────────
supabase==2.9.1

# ─── Environment ─────────────────────────────────────────────
python-dotenv==1.0.1

# ─── Job scraping ────────────────────────────────────────────
python-jobspy==1.1.82

# ─── AI / Embeddings ─────────────────────────────────────────
cohere==5.11.0

# ─── Numerics ────────────────────────────────────────────────
numpy==1.26.3

# ─── Email ───────────────────────────────────────────────────
resend==2.3.0

# ─── Auth / JWT ──────────────────────────────────────────────
pyjwt==2.8.0
cryptography==43.0.3
